grpc_server/requirements.txt
requests==2.18.4
grpcio-tools==1.36.1